Diary broken down into components
